TS-5: refactoring toggl sync handlers
